[PROTO-1411] Add standalone uptime check container with BoltDB
