+++ dev-tools/compose/docker-compose.uptime.yml
@@ -1,0 +1,26 @@
+version: '3.9'
+
+services:
+  uptime:
+    build:
+      context: ${PROJECT_ROOT}/monitoring/uptime
+      dockerfile: Dockerfile
+    restart: unless-stopped
+    environment:
+      - dirTemplate=/data/bolt_%d
+      - hostNameTemplate=http://audius-protocol-creator-node-%d
+    env_file: .env
+    ports:
+      - 1996:1996
+    volumes:
+      - mediorum:/data
+    healthcheck:
+      test:
+        [
+          'CMD-SHELL',
+          'curl --fail http://localhost:1996/health_check || exit 1'
+        ]
+      interval: 10s
+      start_period: 60s
+      timeout: 5s
+      retries: 20
