--- monitoring/peer_health/registrar/the_graph_test.go
@@ -1,21 +1,0 @@
-package registrar
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestTheGraph(t *testing.T) {
-	t.Skip()
-
-	{
-		p := NewGraphProd()
-		fmt.Println(p.Peers("all"))
-	}
-
-	{
-		p := NewGraphStaging()
-		fmt.Println(p.Peers("all"))
-	}
-
-}
