+++ monitoring/uptime/httputil/httputil.go
@@ -1,0 +1,11 @@
+package httputil
+
+import "regexp"
+
+var (
+	trailingSlash = regexp.MustCompile("/$")
+)
+
+func RemoveTrailingSlash(s string) string {
+	return trailingSlash.ReplaceAllString(s, "")
+}
